pyproject.toml
[project]
name = "shiyi"
version = "2.0.0"
description = "Shiyi — 私人定制智能助手，可部署于任意主机"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # OpenAI协议客户端
    "openai>=1.12.0",
    # TTS
    "edge-tts>=6.1.0",
    # 配置管理
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    # 日志
    "loguru>=0.7.2",
    # 数据库
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "kuzu>=0.6.0",
    # Web API
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    # 异步工具
    "aiofiles>=23.2.0",
    "ddgs>=9.10.0",
    "httpx>=0.28.1",
    # 搜索
    "tavily-python>=0.5.0",
    # TUI
    "textual>=0.80.0",
]

[project.scripts]
shiyi = "main:run"

[project.optional-dependencies]
voice = [
    # 音频处理
    "pyaudio>=0.2.14",
    "numpy<2",
    # 唤醒词检测
    "openwakeword>=0.5.0",
    # VAD语音活动检测
    "torch==2.1.2",
    "torchaudio==2.1.2",
    # 腾讯云SDK
    "tencentcloud-sdk-python>=3.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.hatch.build.targets.wheel]
packages = ["audio", "config", "core", "engines", "utils"]
